# docs/public/openapi.yaml
openapi: 3.0.3
info:
  title: pici API
  version: 0.1.0
  description: |
    The pici HTTP API. It drives projects, variables/secrets, and workflow
    executions. Day-to-day operations are usually done through the
    [CLI](/guide/cli) (`pici-cli`), but every operation is available over HTTP.

    ## Authentication

    The API token is required (set via `-api-token` or `PICI_API_TOKEN`).
    Send it either as `Authorization: Bearer <token>` or `X-API-Token: <token>`.

    `/health`, `/api/webhooks/*` and `/badge/*` are exempt (webhooks use their
    own signature/token, badges are meant to be embedded publicly).

    ## Path parameters

    `{id}` on project routes accepts either the project **UUID** or its **name**.
    Execution routes are nested under a project and use `{executionID}`, a
    per-project auto-incrementing number.
servers:
  - url: "{scheme}://{host}:{port}"
    description: pici server (editable from Swagger UI)
    variables:
      scheme:
        default: http
        enum: [http, https]
        description: Protocol
      host:
        default: localhost
        description: Host where pici is running
      port:
        default: "8080"
        description: HTTP listen port (`-http-addr`)
security:
  - BearerAuth: []
  - ApiToken: []
...
paths:
...
  /api/projects/{id}/executions:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
    post:
      tags: [Executions]
      summary: Trigger a run
      description: |
        Enqueues a workflow execution. `ref` defaults to the project's default
        branch. Send `application/json` to run the repo as cloned, or
        `multipart/form-data` to run an uploaded worktree snapshot (a `.tar.gz`
        of the repo, `.git/` included but without `.git/config`). Snapshot
        executions are marked `source: snapshot`, cannot be rebuilt, and do not
        post check runs/commit statuses.
      operationId: createExecution
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExecutionRequest"
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/SnapshotExecutionRequest"
      responses:
        "202":
          description: Execution enqueued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Execution"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "413":
          description: Uploaded snapshot exceeds the maximum size (`-max-snapshot-size`)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/InternalError"
    get:
      tags: [Executions]
      summary: List executions
      operationId: listExecutions
      parameters:
        - name: limit
          in: query
          description: Maximum number of executions to return
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of executions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Execution"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/projects/{id}/executions/{executionID}:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - $ref: "#/components/parameters/ExecutionId"
    get:
      tags: [Executions]
      summary: Get an execution
      operationId: getExecution
      responses:
        "200":
          description: Execution found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Execution"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: "Send the API token as `Authorization: Bearer <token>`."
    ApiToken:
      type: apiKey
      in: header
      name: X-API-Token
      description: Send the API token in the `X-API-Token` header.
  parameters:
    ProjectId:
      name: id
      in: path
      required: true
      description: Project ID (UUID) or name
      schema:
        type: string
    ExecutionId:
      name: executionID
      in: path
      required: true
      description: Per-project execution number
      schema:
        type: integer
        format: int64
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    Status:
      type: string
      enum: [pending, running, success, failed, canceled]
    Trigger:
      type: string
      enum: [manual, webhook, cron, rebuild]
    Source:
      type: string
      enum: [git, snapshot]
      description: Whether the code was cloned (`git`) or uploaded (`snapshot`).
    StepStatus:
      type: string
      enum: [pending, running, success, failed, skipped, canceled]
...
    Error:
      type: object
      properties:
        error:
          type: string
...
    StepResult:
      type: object
      required: [name, status, exit_code]
      properties:
        name:
          type: string
        status:
          $ref: "#/components/schemas/StepStatus"
        started_at:
          type: string
          format: date-time
          nullable: true
        finished_at:
          type: string
          format: date-time
          nullable: true
        exit_code:
          type: integer
        error:
          type: string
        env:
          type: array
          items:
            type: string
          description: Environment variables the step ran with (secrets masked).
    Execution:
      type: object
      required:
        [
          id,
          project_id,
          workflow,
          ref,
          commit_sha,
          status,
          trigger,
          source,
          created_at,
        ]
      properties:
        id:
          type: integer
          format: int64
        project_id:
          type: string
          format: uuid
        project:
          type: string
        workflow:
          type: string
        ref:
          type: string
        commit_sha:
          type: string
        status:
          $ref: "#/components/schemas/Status"
        trigger:
          $ref: "#/components/schemas/Trigger"
        source:
          $ref: "#/components/schemas/Source"
        steps:
          type: array
          items:
            $ref: "#/components/schemas/StepResult"
        error:
          type: string
        started_at:
          type: string
          format: date-time
          nullable: true
        setup_finished_at:
          type: string
          format: date-time
          nullable: true
        finished_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
    ExecutionRequest:
      type: object
      required: [workflow]
      properties:
        workflow:
          type: string
        ref:
          type: string
          description: Branch/tag/SHA. Defaults to the project's default branch.
    SnapshotExecutionRequest:
      type: object
      required: [workflow, snapshot]
      properties:
        workflow:
          type: string
        ref:
          type: string
          description: Branch/tag/SHA the worktree is based on. Defaults to the project's default branch.
        snapshot:
          type: string
          format: binary
          description: A `.tar.gz` archive of the worktree (`.git/` included, `.git/config` excluded).
...